TGL kmeans with 'tidy' output

TGL kmeans with 'tidy' output

TGL_kmeans_tidy(
  df,
  k,
  metric = "euclid",
  max_iter = 40,
  min_delta = 0.0001,
  verbose = FALSE,
  keep_log = FALSE,
  id_column = TRUE,
  reorder_func = "hclust",
  add_to_data = FALSE,
  hclust_intra_clusters = FALSE,
  seed = NULL,
  parallel = getOption("tglkmeans.parallel")
)

Arguments

df

data frame. Each row is a single observation and each column is a dimension. the first column can contain id for each observation (if id_column is TRUE).

k

number of clusters

metric

distance metric for kmeans++ seeding. can be 'euclid', 'pearson' or 'spearman'

max_iter

maximal number of iterations

min_delta

minimal change in assignments (fraction out of all observations) to continue iterating

verbose

display algorithm messages

keep_log

keep algorithm messages in 'log' field

id_column

df's first column contains the observation id

reorder_func

function to reorder the clusters. operates on each center and orders by the result. e.g. reorder_func = mean would calculate the mean of each center and then would reorder the clusters accordingly. If reorder_func = hclust the centers would be ordered by hclust of the euclidean distance of the correlation matrix, i.e. hclust(dist(cor(t(centers)))) if NULL, no reordering would be done.

add_to_data

return also the original data frame with an extra 'clust' column with the cluster ids ('id' is the first column)

hclust_intra_clusters

run hierarchical clustering within each cluster and return an ordering of the observations.

seed

seed for the c++ random number generator

parallel

cluster every cluster parallelly (if hclust_intra_clusters is true)

Value

list with the following components:

cluster:

tibble with `id` column with the observation id (`1:n` if no id column was supplied), and `clust` column with the observation assigned cluster.

centers:

tibble with `clust` column and the cluster centers.

size:

tibble with `clust` column and `n` column with the number of points in each cluster.

data:

tibble with `clust` column the original data frame.

log:

messages from the algorithm run (only if id_column = TRUE).

order:

tibble with 'id' column, 'clust' column, 'order' column with a new ordering if the observations and 'intra_clust_order' column with the order within each cluster. (only if hclust_intra_clusters = TRUE)

See also

Examples

# \dontshow{
tglkmeans.set_parallel(1)
# }

# create 5 clusters normally distributed around 1:5
d <- simulate_data(n = 100, sd = 0.3, nclust = 5, dims = 2, add_true_clust = FALSE)
head(d)
#> # A tibble: 6 x 3
#>   id        V1        V2
#> 1  1 1.0158431 1.3418415
#> 2  2 0.3320804 0.5597070
#> 3  3 0.8628006 0.3859954
#> 4  4 0.8040494 1.2510726
#> 5  5 0.3104177 0.9823126
#> 6  6 1.4977509 0.8090325

# cluster
km <- TGL_kmeans_tidy(d, k = 5, "euclid", verbose = TRUE)
#> id column: id
#> KMEans: will generate seeds
#> KMeans into generate seeds
#> at seed 0
#> add new core from 293 to 0
#> at seed 1
#> done update min distance
#> seed range 350 450
#> picked up 86 dist was 1.57086
#> add new core from 86 to 1
#> at seed 2
#> done update min distance
#> seed range 300 400
#> picked up 417 dist was 1.52982
#> add new core from 417 to 2
#> at seed 3
#> done update min distance
#> seed range 250 350
#> picked up 171 dist was 0.719498
#> add new core from 171 to 3
#> at seed 4
#> done update min distance
#> seed range 200 300
#> picked up 84 dist was 0.470699
#> add new core from 84 to 4
#> KMEans: reassign after init
#> KMEans: iter 0
#> KMEans: iter 1 changed 8
#> KMEans: iter 1
#> KMEans: iter 2 changed 13
#> KMEans: iter 2
#> KMEans: iter 3 changed 15
#> KMEans: iter 3
#> KMEans: iter 4 changed 8
#> KMEans: iter 4
#> KMEans: iter 5 changed 3
#> KMEans: iter 5
#> KMEans: iter 6 changed 1
#> KMEans: iter 6
#> KMEans: iter 7 changed 2
#> KMEans: iter 7
#> KMEans: iter 8 changed 4
#> KMEans: iter 8
#> KMEans: iter 9 changed 8
#> KMEans: iter 9
#> KMEans: iter 10 changed 11
#> KMEans: iter 10
#> KMEans: iter 11 changed 24
#> KMEans: iter 11
#> KMEans: iter 12 changed 24
#> KMEans: iter 12
#> KMEans: iter 13 changed 36
#> KMEans: iter 13
#> KMEans: iter 14 changed 55
#> KMEans: iter 14
#> KMEans: iter 15 changed 27
#> KMEans: iter 15
#> KMEans: iter 16 changed 7
#> KMEans: iter 16
#> KMEans: iter 17 changed 2
#> KMEans: iter 17
#> KMEans: iter 18 changed 0
km
#> $centers
#> # A tibble: 5 x 3
#>   clust        V1        V2
#> 1     1 2.0249572 1.9881035
#> 2     2 4.0749803 4.0160012
#> 3     3 0.9989796 0.9307694
#> 4     4 5.0027680 5.0353432
#> 5     5 3.0264306 3.0947533
#> 
#> $cluster
#> # A tibble: 500 x 2
#>   id clust
#> 1  1     3
#> 2  2     3
#> 3  3     3
#> 4  4     3
#> 5  5     3
#> 6  6     3
#> # ... with 494 more rows
#> 
#> $size
#> # A tibble: 5 x 2
#>   clust   n
#> 1     1 100
#> 2     2  99
#> 3     3 100
#> 4     4 100
#> 5     5 101
#>