Bootstrap clustering

Bootstrap clustering

bootclust(df, N_boot, boot_ratio = 0.75, k_boot = NULL,
  bootstrap_func = "bootstrap_kmeans", add_scores = TRUE, ks = NULL,
  max_k = NULL, parallel = getOption("tglkmeans.parallel"), ...)

Arguments

df
data frame. Each row is a single observation and each column is a dimension. Input to bootstrap_func (first paramter).

N_boot
number of bootstraps.

boot_ratio
fraction of observations to sample in each bootstrap.

k_boot
k to use in bootstrap_func(k parameter).

bootstrap_func
function to use bootstrapping. Should take df as the first parameter, has the following parameters: k, boot_ratio, N_boot, and return a list with coclust, num_trials and coclust_frac.

add_scores
calculate densities of scores for different choices of k.

ks
values of k to calculate scores

max_k
maximal k to test if no ks was given. if NULL, would be chosen as floor(nrow(df) / 40)

parallel
calcualte scores parallely

Value

list with the following components:

mat:

original df. NxM

coclust:

xN matrix with number of times item i was sampled with item j in the same cluster.

num_trials:

NxN matrix with number of times item i was sampled with item j.

coclust_frac:

NxN matrix with coclust / num_trials.

cm:

NxN correlation matrix.

hc:

hierachical clustering of coclust_frac matrix.

coclust_score:

(if add_scores == TRUE) tibble with score for each cluster in for each value of ks.

Examples

data <- simulate_data(n = 200, sd = 0.6, nclust = 5, dims = 2, add_true_clust = TRUE)
bt <- bootclust(data %>% select(id, starts_with("V")), k_boot = 5, N_boot = 100, boot_ratio = 0.75)

#> bootstrapping

#> clustering bootstrapping results

#> calculating score for each k

bt$mat

#> # A tibble: 1,000 x 3
#>       id    V1    V2
#>    <int> <dbl> <dbl>
#>  1     1 0.358 2.05 
#>  2     2 0.669 1.09 
#>  3     3 1.03  1.39 
#>  4     4 0.824 0.981
#>  5     5 0.387 0.761
#>  6     6 0.544 1.50 
#>  7     7 1.68  1.62 
#>  8     8 2.67  0.965
#>  9     9 1.17  0.699
#> 10    10 1.02  1.03 
#> # … with 990 more rows

bt$coclust[1:5, 1:5]

#>    1  2  3  4  5
#> 1 83 59 65 57 53
#> 2 59 74 62 56 45
#> 3 65 62 81 58 54
#> 4 57 56 58 74 49
#> 5 53 45 54 49 65

bt$num_trials[1:5, 1:5]

#>    1  2  3  4  5
#> 1 83 63 69 62 55
#> 2 63 74 62 57 46
#> 3 69 62 81 58 54
#> 4 62 57 58 74 49
#> 5 55 46 54 49 65

bt$coclust_frac[1:5, 1:5]

#>           1         2        3         4         5
#> 1 1.0000000 0.9365079 0.942029 0.9193548 0.9636364
#> 2 0.9365079 1.0000000 1.000000 0.9824561 0.9782609
#> 3 0.9420290 1.0000000 1.000000 1.0000000 1.0000000
#> 4 0.9193548 0.9824561 1.000000 1.0000000 1.0000000
#> 5 0.9636364 0.9782609 1.000000 1.0000000 1.0000000

bt$cm[1:5, 1:5]

#>           1         2         3         4         5
#> 1 1.0000000 0.8817969 0.8248100 0.8264839 0.8324510
#> 2 0.8817969 1.0000000 0.9250820 0.8838143 0.8908291
#> 3 0.8248100 0.9250820 1.0000000 0.9613295 0.9688622
#> 4 0.8264839 0.8838143 0.9613295 1.0000000 0.9933702
#> 5 0.8324510 0.8908291 0.9688622 0.9933702 1.0000000

bt$hc

#> 
#> Call:
#> hclust(d = tgstat::tgs_dist(cm), method = "ward.D2")
#> 
#> Cluster method   : ward.D2 
#> Distance         : euclidian 
#> Number of objects: 1000 
#> 

bt$coclust_score

#> # A tibble: 24,000 x 4
#>    k     i     clust score
#>    <fct> <chr> <int> <dbl>
#>  1 2     1         1 1.000
#>  2 2     2         1 1    
#>  3 2     3         1 1    
#>  4 2     4         1 1    
#>  5 2     5         1 1    
#>  6 2     6         1 1    
#>  7 2     7         1 0.956
#>  8 2     8         1 0.955
#>  9 2     9         1 1    
#> 10 2     10        1 1    
#> # … with 23,990 more rows