Bootstrapping

Bootstrapping

bootstrap_kmeans(df, k, N_boot, boot_ratio = 0.75,
  parallel = getOption("tglkmeans.parallel"), id_column = TRUE,
  seed = NULL, ...)

Arguments

df
data frame. Each row is a single observation and each column is a dimension. the first column can contain id for each observation (if id_column is TRUE).

k
number of clusters

N_boot
number of bootstrapping iterations

boot_ratio
percent of observations to sample on each iteration

parallel
run parallely (using doMC backend)

id_column
df's first column contains the observation id

seed
seed for the c++ random number generator

...
additional parameters to TGL_kmeans

Value

list with the following components:

coclust:

NxN matrix (where N is the number of observations) with the number of times observation i and j occured in the same cluster.

num_trials:

NxN matrix with the number of times observation i and j where sampled together.

coclust_frac:

fraction of times observation i and j where clustered together out of the times they were sampled together (coclust matrix divided by num_trails matrix).

.

Examples

# NOT RUN {
library(dplyr)
d <- simulate_data(nclust = 6)
bootstrap <- bootstrap_kmeans(d, k = 6, N_boot = 100)
names(bootstrap)
bootstrap$coclust[1:5, 1:5]
bootstrap$num_trials[1:5, 1:5]
bootstrap$coclust_frac[1:5, 1:5]

bootstrap_kmeans(d, k = 6, N_boot = 100, tidy = TRUE)
# }