Cutree of bootstrap clustering

Cutree of bootstrap clustering

cutree_bootclust(bt, k, min_coclust = 0.5, tidy = FALSE)

Arguments

bt
output of bootclust

k
an integer scalar or vector with the desired number of groups

min_coclust
minimal co-clust score for observation in a cluster

tidy
tidy output

Value

if tidy == TRUE: bt with the following aditional fields:

clust:

tibble with `id` column with the observation id (`1:n` if no id column was supplied), and `clust` column with the observation assigned cluster.

centers:

tibble with `clust` column and the cluster centers.

size:

tibble with `clust` column and `n` column with the number of points in each cluster.

score:

if tidy == FALSE: bt with the following aditional fields:

cluster:

A vector of integers (from ‘1:k’) indicating the cluster to which each point is allocated.

centers:

A matrix of cluster centres.

size:

The number of points in each cluster.

score:

Examples

data <- simulate_data(n = 200, sd = 0.6, nclust = 5, dims = 2, add_true_clust = TRUE)
bt <- bootclust(data %>% select(id, starts_with("V")), k_boot = 5, N_boot = 100, boot_ratio = 0.75)

#> bootstrapping

#> clustering bootstrapping results

#> calculating score for each k

names(bt)

#> [1] "mat"           "coclust"       "num_trials"    "coclust_frac" 
#> [5] "cm"            "hc"            "coclust_score"

# tidy output
bt_tidy <- cutree_bootclust(bt, k_boot = 5, min_coclust = 0.6, tidy = TRUE)

#> Error in cutree_bootclust(bt, k_boot = 5, min_coclust = 0.6, tidy = TRUE): unused argument (k_boot = 5)

names(bt_tidy)

#> Error in eval(expr, envir, enclos): object 'bt_tidy' not found

bt_tidy$clust

#> Error in eval(expr, envir, enclos): object 'bt_tidy' not found

bt_tidy$centers

#> Error in eval(expr, envir, enclos): object 'bt_tidy' not found

bt_tidy$size

#> Error in eval(expr, envir, enclos): object 'bt_tidy' not found

bt_tidy$score

#> Error in eval(expr, envir, enclos): object 'bt_tidy' not found

# non tidy output
bt <- cutree_bootclust(bt, k_boot = 5, min_coclust = 0.6, tidy = FALSE)

#> Error in cutree_bootclust(bt, k_boot = 5, min_coclust = 0.6, tidy = FALSE): unused argument (k_boot = 5)

bt$cluster[1:5]

#> NULL

bt$centers[1:5, ]

#> NULL

bt$size

#> NULL

bt$score

#> NULL