kmeans++ with return value similar to R kmeans

kmeans++ with return value similar to R kmeans

TGL_kmeans(
  df,
  k,
  metric = "euclid",
  max_iter = 40,
  min_delta = 0.0001,
  verbose = FALSE,
  keep_log = FALSE,
  id_column = TRUE,
  reorder_func = "hclust",
  hclust_intra_clusters = FALSE,
  seed = NULL,
  parallel = getOption("tglkmeans.parallel")
)

Arguments

df
data frame. Each row is a single observation and each column is a dimension. the first column can contain id for each observation (if id_column is TRUE).

k
number of clusters

metric
distance metric for kmeans++ seeding. can be 'euclid', 'pearson' or 'spearman'

max_iter
maximal number of iterations

min_delta
minimal change in assignments (fraction out of all observations) to continue iterating

verbose
display algorithm messages

keep_log
keep algorithm messages in 'log' field

id_column
df's first column contains the observation id

reorder_func
function to reorder the clusters. operates on each center and orders by the result. e.g. reorder_func = mean would calculate the mean of each center and then would reorder the clusters accordingly. If reorder_func = hclust the centers would be ordered by hclust of the euclidian distance of the correlation matrix, i.e. hclust(dist(cor(t(centers)))) if NULL, no reordering would be done.

hclust_intra_clusters
run hierarchical clustering within each cluster and return an ordering of the observations.

seed
seed for the c++ random number generator

parallel
cluster every cluster parallely (if hclust_intra_clusters is true)

Value

list with the following components:

cluster:

A vector of integers (from ‘1:k’) indicating the cluster to which each point is allocated.

centers:

A matrix of cluster centers.

size:

The number of points in each cluster.

log:

messages from the algorithm run (only if id_column == TRUE).

order:

A vector of integers with the new ordering if the observations. (only if hclust_intra_clusters = TRUE)

See also

Examples

# create 5 clusters normally distributed around 1:5
d <- simulate_data(n = 100, sd = 0.3, nclust = 5, dims = 2, add_true_clust = FALSE)
head(d)

#> # A tibble: 6 × 3
#>      id    V1    V2
#>   <int> <dbl> <dbl>
#> 1     1 0.580 0.884
#> 2     2 1.08  0.764
#> 3     3 0.269 0.683
#> 4     4 0.998 0.761
#> 5     5 1.19  0.473
#> 6     6 1.34  0.793

# cluster
km <- TGL_kmeans(d, k = 5, "euclid", verbose = TRUE)

#> id column: id

#> KMEans: will generate seeds
#> KMeans into generate seeds
#> at seed 0
#> add new core from 267 to 0
#> at seed 1
#> done update min distance
#> seed range 350 450
#> picked up 2 dist was 1.72777
#> add new core from 2 to 1
#> at seed 2
#> done update min distance
#> seed range 300 400
#> picked up 473 dist was 1.47502
#> add new core from 473 to 2
#> at seed 3
#> done update min distance
#> seed range 250 350
#> picked up 307 dist was 0.661523
#> add new core from 307 to 3
#> at seed 4
#> done update min distance
#> seed range 200 300
#> picked up 168 dist was 0.557587
#> add new core from 168 to 4
#> KMEans: reassign after init
#> KMEans: iter 0
#> KMEans: iter 1 changed 3
#> KMEans: iter 1
#> KMEans: iter 2 changed 0

names(km)

#> [1] "cluster" "centers" "size"   

km$centers

#>            V1       V2
#> [1,] 2.000776 2.053331
#> [2,] 3.969391 4.038842
#> [3,] 4.975310 5.048907
#> [4,] 2.964869 3.008410
#> [5,] 1.019032 1.034553

head(km$cluster)

#> 1 2 3 4 5 6 
#> 5 5 5 5 5 5 

km$size

#>   1   2   3   4   5 
#> 101  98  99 102 100