kmeans++ with return value similar to R kmeans

kmeans++ with return value similar to R kmeans

TGL_kmeans(df, k, metric = "euclid", max_iter = 40,
  min_delta = 1e-04, verbose = FALSE, keep_log = FALSE,
  id_column = TRUE, reorder_func = "hclust",
  hclust_intra_clusters = FALSE, seed = NULL,
  parallel = getOption("tglkmeans.parallel"))

Arguments

df
data frame. Each row is a single observation and each column is a dimension. the first column can contain id for each observation (if id_column is TRUE).

k
number of clusters

metric
distance metric for kmeans++ seeding. can be 'euclid', 'pearson' or 'spearman'

max_iter
maximal number of iterations

min_delta
minimal change in assignments (fraction out of all observations) to continue iterating

verbose
display algorithm messages

keep_log
keep algorithm messages in 'log' field

id_column
df's first column contains the observation id

reorder_func
function to reorder the clusters. operates on each center and orders by the result. e.g. reorder_func = mean would calculate the mean of each center and then would reorder the clusters accordingly. If reorder_func = hclust the centers would be ordered by hclust of the euclidian distance of the corelation matrix, i.e. hclust(dist(cor(t(centers)))) if NULL, no reordering would be done.

hclust_intra_clusters
run hierarchical clustering within each cluster and return an ordering of the observations.

seed
seed for the c++ random number generator

parallel
cluster every cluster parallely (if hclust_intra_clusters is true)

Value

list with the following components:

cluster:

A vector of integers (from ‘1:k’) indicating the cluster to which each point is allocated.

centers:

A matrix of cluster centres.

size:

The number of points in each cluster.

log:

messages from the algorithm run (only if id_column == TRUE).

order:

A vector of integers with the new ordering if the observations. (only if hclust_intra_clusters = TRUE)

See also

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# create 5 clusters normally distributed around 1:5
d <- simulate_data(n = 100, sd = 0.3, nclust = 5, dims = 2, add_true_clust = FALSE)
head(d)

#> # A tibble: 6 x 3
#>      id    V1    V2
#>   <int> <dbl> <dbl>
#> 1     1 0.580 0.884
#> 2     2 1.08  0.764
#> 3     3 0.269 0.683
#> 4     4 0.998 0.761
#> 5     5 1.19  0.473
#> 6     6 1.34  0.793

# cluster
km <- TGL_kmeans(d, k = 5, "euclid", verbose = TRUE)

#> id column: id

#> KMEans: will generate seeds
#> KMeans into generate seeds
#> at seed 0
#> add new core from 26 to 0
#> at seed 1
#> done update min distance
#> seed range 350 450
#> picked up 481 dist was 2.96102
#> add new core from 481 to 1
#> at seed 2
#> done update min distance
#> seed range 300 400
#> picked up 206 dist was 1.35296
#> add new core from 206 to 2
#> at seed 3
#> done update min distance
#> seed range 250 350
#> picked up 128 dist was 0.672143
#> add new core from 128 to 3
#> at seed 4
#> done update min distance
#> seed range 200 300
#> picked up 352 dist was 0.651756
#> add new core from 352 to 4
#> KMEans: reassign after init
#> KMEans: iter 0
#> KMEans: iter 1 changed 10
#> KMEans: iter 1
#> KMEans: iter 2 changed 1
#> KMEans: iter 2
#> KMEans: iter 3 changed 0

names(km)

#> [1] "cluster" "centers" "size"   

km$centers

#>            V1       V2
#> [1,] 3.972232 4.046328
#> [2,] 2.004643 2.058945
#> [3,] 2.977600 3.015143
#> [4,] 1.019032 1.034553
#> [5,] 4.975310 5.048907

head(km$cluster)

#> 1 2 3 4 5 6 
#> 4 4 4 4 4 4 

km$size

#>   1   2   3   4   5 
#>  97 102 102 100  99