kmeans++ with return value similar to R kmeans

kmeans++ with return value similar to R kmeans

TGL_kmeans(
  df,
  k,
  metric = "euclid",
  max_iter = 40,
  min_delta = 0.0001,
  verbose = FALSE,
  keep_log = FALSE,
  id_column = TRUE,
  reorder_func = "hclust",
  hclust_intra_clusters = FALSE,
  seed = NULL,
  parallel = getOption("tglkmeans.parallel")
)

Arguments

df

data frame. Each row is a single observation and each column is a dimension. the first column can contain id for each observation (if id_column is TRUE).

k

number of clusters

metric

distance metric for kmeans++ seeding. can be 'euclid', 'pearson' or 'spearman'

max_iter

maximal number of iterations

min_delta

minimal change in assignments (fraction out of all observations) to continue iterating

verbose

display algorithm messages

keep_log

keep algorithm messages in 'log' field

id_column

df's first column contains the observation id

reorder_func

function to reorder the clusters. operates on each center and orders by the result. e.g. reorder_func = mean would calculate the mean of each center and then would reorder the clusters accordingly. If reorder_func = hclust the centers would be ordered by hclust of the euclidean distance of the correlation matrix, i.e. hclust(dist(cor(t(centers)))) if NULL, no reordering would be done.

hclust_intra_clusters

run hierarchical clustering within each cluster and return an ordering of the observations.

seed

seed for the c++ random number generator

parallel

cluster every cluster parallelly (if hclust_intra_clusters is true)

Value

list with the following components:

cluster:

A vector of integers (from ‘1:k’) indicating the cluster to which each point is allocated.

centers:

A matrix of cluster centers.

size:

The number of points in each cluster.

log:

messages from the algorithm run (only if id_column == TRUE).

order:

A vector of integers with the new ordering if the observations. (only if hclust_intra_clusters = TRUE)

See also

Examples

# \dontshow{
tglkmeans.set_parallel(1)
# }

# create 5 clusters normally distributed around 1:5
d <- simulate_data(n = 100, sd = 0.3, nclust = 5, dims = 2, add_true_clust = FALSE)
head(d)
#> # A tibble: 6 x 3
#>   id        V1        V2
#> 1  1 0.9948935 0.9522707
#> 2  2 0.6219252 0.5360106
#> 3  3 0.6422511 1.2970306
#> 4  4 1.1528729 0.5651336
#> 5  5 0.7468167 0.8703051
#> 6  6 0.6647931 0.7234569

# cluster
km <- TGL_kmeans(d, k = 5, "euclid", verbose = TRUE)
#> id column: id
#> KMEans: will generate seeds
#> KMeans into generate seeds
#> at seed 0
#> add new core from 384 to 0
#> at seed 1
#> done update min distance
#> seed range 350 450
#> picked up 2 dist was 2.2239
#> add new core from 2 to 1
#> at seed 2
#> done update min distance
#> seed range 300 400
#> picked up 284 dist was 0.870011
#> add new core from 284 to 2
#> at seed 3
#> done update min distance
#> seed range 250 350
#> picked up 477 dist was 0.633211
#> add new core from 477 to 3
#> at seed 4
#> done update min distance
#> seed range 200 300
#> picked up 168 dist was 0.55472
#> add new core from 168 to 4
#> KMEans: reassign after init
#> KMEans: iter 0
#> KMEans: iter 1 changed 1
#> KMEans: iter 1
#> KMEans: iter 2 changed 0
names(km)
#> [1] "cluster" "centers" "size"   
km$centers
#>             V1        V2
#> [1,] 1.9628351 2.0191987
#> [2,] 3.9457889 4.0083160
#> [3,] 0.9760559 0.9918235
#> [4,] 2.9723847 2.9043789
#> [5,] 5.0155549 5.0024590
head(km$cluster)
#> 1 2 3 4 5 6 
#> 3 3 3 3 3 3 
km$size
#>   1   2   3   4   5 
#> 102 103  97  99  99