TGL kmeans with 'tidy' output

TGL kmeans with 'tidy' output

TGL_kmeans_tidy(df, k, metric = "euclid", max_iter = 40,
  min_delta = 1e-04, verbose = FALSE, keep_log = TRUE,
  id_column = TRUE, reorder_func = "hclust", add_to_data = FALSE,
  hclust_intra_clusters = FALSE, seed = NULL, bootstrap = FALSE, ...)

Arguments

df
data frame. Each row is a single observation and each column is a dimension. the first column can contain id for each observation (if id_column is TRUE).

k
number of clusters

metric
distance metric for kmeans++ seeding. can be 'euclid', 'pearson' or 'spearman'

max_iter
maximal number of iterations

min_delta
minimal change in assignments (fraction out of all observations) to continue iterating

verbose
display algorithm messages

keep_log
keep algorithm messages in 'log' field

id_column
df's first column contains the observation id

reorder_func
function to reorder the clusters. operates on each center and orders by the result. e.g. reorder_func = mean would calculate the mean of each center and then would reorder the clusters accordingly. If reorder_func = hclust the centers would be ordered by hclust of the euclidian distance of the corelation matrix, i.e. hclust(dist(cor(t(centers)))) if NULL, no reordering would be done.

add_to_data
return also the original data frame with an extra 'clust' column with the cluster ids ('id' is the first column)

hclust_intra_clusters
run hierarchical clustering within each cluster and return an ordering of the observations.

seed
seed for the c++ random number generator

bootstrap
bootstrap to estimate robustness of the clusters

...

...
Arguments passed on to bootstrap_kmeans

df

data frame. Each row is a single observation and each column is a dimension. the first column can contain id for each observation (if id_column is TRUE).

k

number of clusters

N_boot

number of bootstrapping iterations

boot_ratio

percent of observations to sample on each iteration

parallel

run parallely (using doMC backend)

id_column

df's first column contains the observation id

seed

seed for the c++ random number generator

Value

list with the following components:

cluster:

tibble with `id` column with the observation id (`1:n` if no id column was supplied), and `clust` column with the observation assigned cluster.

centers:

tibble with `clust` column and the cluster centers.

size:

tibble with `clust` column and `n` column with the number of points in each cluster.

data:

tibble with `clust` column the original data frame.

log:

messages from the algorithm run (only if id_column = TRUE).

bootstrap:

tibble with 'clust' column and 'robust' column with the number of times the members of the clusters were clustered together divided by the total times they were sampled together. (only if bootstrap = TRUE)

order:

tibble with 'id' column, 'clust' column, 'order' column with a new ordering if the observations and 'intra_clust_order' column with the order within each cluster. (only if hclust_intra_clusters = TRUE)

See also

Examples

library(dplyr)
# create 5 clusters normally distributed around 1:5
d <- simulate_data(n = 100, sd = 0.3, nclust = 5, dims = 2, add_true_clust = FALSE)
head(d)

#> # A tibble: 6 x 3
#>      id      V1    V2
#>   <int>   <dbl> <dbl>
#> 1     1  0.949  0.764
#> 2     2  0.807  0.841
#> 3     3  0.776  0.666
#> 4     4  0.731  0.822
#> 5     5  1.59   0.832
#> 6     6 -0.0535 1.75 

# cluster
km <- TGL_kmeans_tidy(d, k = 5, "euclid", verbose = TRUE)

#> id column: id

#> KMEans: will generate seeds
#> KMeans into generate seeds
#> at seed 0
#> add new core from 407 to 0
#> at seed 1
#> done update min distance
#> seed range 350 450
#> picked up 0 dist was 2.94227
#> add new core from 0 to 1
#> at seed 2
#> done update min distance
#> seed range 300 400
#> picked up 266 dist was 1.32034
#> add new core from 266 to 2
#> at seed 3
#> done update min distance
#> seed range 250 350
#> picked up 357 dist was 0.648405
#> add new core from 357 to 3
#> at seed 4
#> done update min distance
#> seed range 200 300
#> picked up 156 dist was 0.716407
#> add new core from 156 to 4
#> KMEans: reassign after init
#> KMEans: iter 0
#> KMEans: iter 1 changed 4
#> KMEans: iter 1
#> KMEans: iter 2 changed 1
#> KMEans: iter 2
#> KMEans: iter 3 changed 0

km

#> $centers
#> # A tibble: 5 x 3
#>   clust    V1    V2
#>   <int> <dbl> <dbl>
#> 1     1 2.97  2.98 
#> 2     2 4.98  4.98 
#> 3     3 0.975 0.998
#> 4     4 4.02  4.01 
#> 5     5 2.05  1.98 
#> 
#> $cluster
#> # A tibble: 500 x 2
#>    id    clust
#>    <chr> <int>
#>  1 1         3
#>  2 2         3
#>  3 3         3
#>  4 4         3
#>  5 5         3
#>  6 6         3
#>  7 7         3
#>  8 8         3
#>  9 9         3
#> 10 10        3
#> # … with 490 more rows
#> 
#> $size
#> # A tibble: 5 x 2
#>   clust     n
#>   <int> <int>
#> 1     1    99
#> 2     2   102
#> 3     3   101
#> 4     4    99
#> 5     5    99
#> 

# bootstrapping
km <- TGL_kmeans_tidy(d, k = 5, "euclid", N_boot = 100, bootstrap = TRUE)

#> bootstrapping

km$bootstrap

#> # A tibble: 5 x 2
#> # Groups:   clust [5]
#>   clust robust
#>   <int>  <dbl>
#> 1     1  0.979
#> 2     2  0.976
#> 3     3  0.985
#> 4     4  0.985
#> 5     5  0.986