TGL kmeans with 'tidy' output

TGL kmeans with 'tidy' output

TGL_kmeans_tidy(df, k, metric = "euclid", max_iter = 40,
  min_delta = 1e-04, verbose = FALSE, keep_log = FALSE,
  id_column = TRUE, reorder_func = "hclust", add_to_data = FALSE,
  hclust_intra_clusters = FALSE, seed = NULL,
  parallel = getOption("tglkmeans.parallel"))

Arguments

df
data frame. Each row is a single observation and each column is a dimension. the first column can contain id for each observation (if id_column is TRUE).

k
number of clusters

metric
distance metric for kmeans++ seeding. can be 'euclid', 'pearson' or 'spearman'

max_iter
maximal number of iterations

min_delta
minimal change in assignments (fraction out of all observations) to continue iterating

verbose
display algorithm messages

keep_log
keep algorithm messages in 'log' field

id_column
df's first column contains the observation id

reorder_func
function to reorder the clusters. operates on each center and orders by the result. e.g. reorder_func = mean would calculate the mean of each center and then would reorder the clusters accordingly. If reorder_func = hclust the centers would be ordered by hclust of the euclidian distance of the corelation matrix, i.e. hclust(dist(cor(t(centers)))) if NULL, no reordering would be done.

add_to_data
return also the original data frame with an extra 'clust' column with the cluster ids ('id' is the first column)

hclust_intra_clusters
run hierarchical clustering within each cluster and return an ordering of the observations.

seed
seed for the c++ random number generator

parallel
cluster every cluster parallely (if hclust_intra_clusters is true)

Value

list with the following components:

cluster:

tibble with `id` column with the observation id (`1:n` if no id column was supplied), and `clust` column with the observation assigned cluster.

centers:

tibble with `clust` column and the cluster centers.

size:

tibble with `clust` column and `n` column with the number of points in each cluster.

data:

tibble with `clust` column the original data frame.

log:

messages from the algorithm run (only if id_column = TRUE).

order:

tibble with 'id' column, 'clust' column, 'order' column with a new ordering if the observations and 'intra_clust_order' column with the order within each cluster. (only if hclust_intra_clusters = TRUE)

See also

Examples

library(dplyr)
# create 5 clusters normally distributed around 1:5
d <- simulate_data(n = 100, sd = 0.3, nclust = 5, dims = 2, add_true_clust = FALSE)
head(d)

#> # A tibble: 6 x 3
#>      id    V1    V2
#>   <int> <dbl> <dbl>
#> 1     1 0.899 0.488
#> 2     2 0.935 0.743
#> 3     3 1.19  0.957
#> 4     4 0.615 0.903
#> 5     5 0.610 0.948
#> 6     6 0.887 0.629

# cluster
km <- TGL_kmeans_tidy(d, k = 5, "euclid", verbose = TRUE)

#> id column: id

#> KMEans: will generate seeds
#> KMeans into generate seeds
#> at seed 0
#> add new core from 229 to 0
#> at seed 1
#> done update min distance
#> seed range 350 450
#> picked up 453 dist was 1.50538
#> add new core from 453 to 1
#> at seed 2
#> done update min distance
#> seed range 300 400
#> picked up 42 dist was 1.36515
#> add new core from 42 to 2
#> at seed 3
#> done update min distance
#> seed range 250 350
#> picked up 331 dist was 0.687862
#> add new core from 331 to 3
#> at seed 4
#> done update min distance
#> seed range 200 300
#> picked up 168 dist was 0.72591
#> add new core from 168 to 4
#> KMEans: reassign after init
#> KMEans: iter 0
#> KMEans: iter 1 changed 4
#> KMEans: iter 1
#> KMEans: iter 2 changed 2
#> KMEans: iter 2
#> KMEans: iter 3 changed 1
#> KMEans: iter 3
#> KMEans: iter 4 changed 0

km

#> $centers
#> # A tibble: 5 x 3
#>   clust    V1    V2
#>   <int> <dbl> <dbl>
#> 1     1 0.945 0.926
#> 2     2 3.03  3.01 
#> 3     3 5.00  4.99 
#> 4     4 4.04  4.04 
#> 5     5 1.96  2.02 
#> 
#> $cluster
#> # A tibble: 500 x 2
#>    id    clust
#>    <chr> <int>
#>  1 1         1
#>  2 2         1
#>  3 3         1
#>  4 4         1
#>  5 5         1
#>  6 6         1
#>  7 7         1
#>  8 8         1
#>  9 9         1
#> 10 10        1
#> # … with 490 more rows
#> 
#> $size
#> # A tibble: 5 x 2
#>   clust     n
#>   <int> <int>
#> 1     1    98
#> 2     2   101
#> 3     3    99
#> 4     4   100
#> 5     5   102
#>